Fix GetBlobs and GetMissingBlobs request/response (#31)
Blobs are not globally addressable across the BSR, largely for SOC2
compliance so that we can safely and cheaply delete all assets for a
repository. Blobs are scoped to a repository.

Therefore, when we query for missing blobs, we have to query for missing
blobs in the context of a `ModuleRef`. I chose to take the route of
querying for multiple sets of module blobs in a single request rather
than one request per set of module blobs.