Adjust page token validation (#34)
We wanted to either pick some arbitrary large value or remove it
entirely. Others pointed out that 255 is short enough that it might be a
bit too small for base64'd cursor pagination, so I picked 4096 as an
arbitrary value.

The justification for this is that _removing_ validation later would be
backwards compatible, but adding validation later would be backwards
incompatible. So a lenient limit seems like a decent idea.

I think this is the last of the validation issues. I made a couple
additional passes to make sure and didn't see anything else that
obviously needed attention.

<!-- Closes BSR-2930. -->