Revamp CreateCommits to just create Commits and add specialized Push and Rebase endpoints (#43)
There are two things going on in this PR:

1. `CreateCommits` is updated so that it simply creates Commits for
Modules without pushing them onto any Branch or assigning any Tags to
them.
- The idea is that Commits can be "uploaded" to the BSR before they are
placed in their correct places in Branch history or assigned Tags. This
is very useful for `sync`, where potentially 1000s of Commits need to be
created before pushing them on a Branch. Because Branch history should
be patched atomically and with a lock, creating Commits in-band becomes
untenable.
- These Commits can only be referenced by their `ID`, `Digest`, or a
`VCSCommitRef` if there was any `AssociatedVCSCommit`.
2. Two new RPCs are added to mutate a `Branch`'s history:
`BranchService.Push` and `BranchService.Rebase`.
- `Push` provides no assertions on the final state of a `Branch`'s
history. This is the model that `buf push` currently uses, where all
concurrent writes win, but the last write is king and becomes the latest
commit.
- `Rebase` allows the caller to assert on the final state of a
`Branch`'s history. Concurrent writes are idempotent as long as they
assert the same final state. If they don't, the first write wins and all
others lose. This is the model that `buf sync` will use.

---------

Co-authored-by: bufdev <bufdev-github@buf.build>
Co-authored-by: Nick Snyder <nsnyder@buf.build>
Co-authored-by: bufdev <4228796+bufdev@users.noreply.github.com>
Co-authored-by: Akshay Shah <akshayjshah@users.noreply.github.com>