Remove VCSCommit and add Commit.source_control_url (#55)
The BSR doesn't need to be in the business of storing source control
information--we don't semantically need a vcs commit hash for example
and we don't care what type of source control is being used.

As a convenience for users, we DO want to store a link back to the
relevant commit in source control so that a user can find relevant info
(e.g. commit description, PR discussion, authors, approvers, etc.).

This change simplifies the data model to only store what we need, which
is a single source control URL per commit.