LabelReviewState -> CommitReviewState and update docs
CommitReviewState is a better name because LabelReviewState seems to imply a label has only one such state. The number of review states is 1:1 with number of commits (not labels).

Update docs to clarify that review state information is returned depending on whether review state was enabled at the time of push for a given commit (not whether it is enabled currently).