Add start_commit_id to ListLabelHistoryRequest (#81)
To view a diff for a commit on a label we need to be able to get the
previous commit on that label. Added `start_commit_id` so we can fetch
the history starting at the desired commit and using a page size of 2
can also get the previous commit.