Add initial Policy specification (#138)
This PR adds support for "check policies" as a feature in our API.

A new entity, `Policy` is added. Policies can only be owned by
organizations,
and are in the same namespace as modules and plugins. A policy has
immutable
versions (`Version`) and each version represents a `Configuration` of
the `Policy`.

Uploading a new `Configuration` of a `Policy` will increment its
`Version`. If the same
`Configuration` is uploaded, then the current latest `Version` is
returned.