Set typed Policy configuration (#146)
This PR sets the Policy configuration to a defined message type,
`PolicyConfig`. This is used on uploading and downloading policies. A
`PolicyConfig` represents the subset of the policy.yaml file that
affects how a Policy is run. This fixes the calculation of the digest to
the contents of the `PolicyConfig`. A digest is no longer affected by
changes to the `policy.yaml` that don't alter the configuration (e.g.
comments and policy name have no change in `PolicyConfig`).