Add disable_builtin field to policy configuration (#153)
This adds the bool field `disable_builtin` to lint and breaking policy
configuration. When set no builtin rules or categories are evaluated.
This is required for use with policies to disable the default rule or
categories being evaluated when not intended. For example, to set only
breaking rules the lint configuration would be set to disable builtin to
ensure the defaults (STANDARD) aren't invoked.